+++ packages/bridge-controller/src/index.ts
@@ -83,8 +83,6 @@ export {
   SortOrder,
   ChainId,
   RequestStatus,
-  BridgeUserAction,
-  BridgeBackgroundAction,
   type TokenFeature,
   type QuoteStreamCompleteData,
   type BridgeControllerGetStateAction,