+++ packages/bridge-status-controller/src/utils/bridge.ts
@@ -2,8 +2,9 @@ import {
   AbortReason,
   FeatureId,
   UnifiedSwapBridgeEventName,
+  BatchSellTradesResponse,
+  RequiredEventContextFromClient,
 } from '@metamask/bridge-controller';
-import type { RequiredEventContextFromClient } from '@metamask/bridge-controller';
 
 import { BridgeStatusControllerMessenger } from '../types';
 
@@ -19,6 +20,12 @@ export const stopPollingForQuotes = (
     // If the trade has a featureId, it means it was submitted outside of the Unified Swap and Bridge experience, so no QuotesReceived event is published
     featureId ? undefined : metricsContext,
   );
+};
+
+export const getBatchSellTrades = (
+  messenger: BridgeStatusControllerMessenger,
+): BatchSellTradesResponse | null => {
+  return messenger.call('BridgeController:getState').batchSellTrades;
 };
 
 export const trackMetricsEvent = ({